+++ spec/action_spec.rb
@@ -1,0 +1,23 @@
+require 'spec_helper'
+
+describe 'action' do
+
+xit "adds a task to a project list" do
+  end
+
+xit "sets project ID to the task it's being added to" do
+  end
+
+xit "adds a task to a project list" do
+  end
+
+xit "puts to the console the the task and the project it's added to" do
+  end
+
+xit "retrieves tasks from a project sorted by priority" do
+  end
+
+xit "retrieves tasks with the same priority sorted by date" do
+  end
+
+end
